Re-factor of Prompt to get data out of main Promp definition
