+++ edsl/exceptions/prompts.py
@@ -1,0 +1,3 @@
+class TemplateRenderError(Exception):
+    "TODO: Move to exceptions file"
+    pass
