+++ edsl/prompts/__init__.py
@@ -1,0 +1,2 @@
+from edsl.prompts.registry import get_classes
+from edsl.prompts.Prompt import Prompt
