+++ tests/prompts/test_Prompts.py
@@ -1,9 +1,9 @@
 import doctest
-import edsl.prompts.Prompt
+import edsl.prompts
 
 
 def test_doctests_in_prompt():
-    doctest.testmod(edsl.prompts.Prompt)
+    doctest.testmod(edsl.prompts)
 
 
 def test_foo():
